Modified the project structure and fixed task package naming
