+++ src/main/java/seedu/duke/command/DoneCommand.java
@@ -1,6 +1,6 @@
 package seedu.duke.command;
 
-import seedu.duke.tasks.TaskList;
+import seedu.duke.list.TaskList;
 import seedu.duke.ui.Ui;
 
 /**
