--- src/main/java/seedu/duke/subject/Subject.java
@@ -1,4 +1,0 @@
-package seedu.duke.subject;
-
-public class Subject {
-}
