+++ src/main/java/seedu/duke/task/Todo.java
@@ -1,15 +1,14 @@
-package seedu.duke.tasks;
+package seedu.duke.task;
 
-public class Deadline extends Task {
+import seedu.duke.list.TaskList;
 
-    protected String by;
+public class Todo extends Task {
 
-    public Deadline(String description,boolean isDone, String by) {
+    public Todo(String description, boolean isDone) {
         super(description,isDone);
-        this.by = by;
     }
 
-    public void printDeadline(TaskList taskList){
+    public void printTodo(TaskList taskList){
         System.out.println("____________________________________________________________\n" +
                 "Got it. I've added this task:\n  " + this + "\n" +
                 "Now you have " + taskList.getList().size() + (taskList.getList().size() == 1
@@ -18,7 +17,7 @@ public class Deadline extends Task {
     }
 
     @Override
-    public String toString() {
-        return "[D]" + super.toString() + " (by: " + by + ")";
+    public String toString(){
+        return "[T]" + super.toString();
     }
 }
